cloud_functions/generate_thumbnail/requirements.txt
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
pillow==10.3.0